move the java core to stryker4jvm.core.model
