+++ stryker4jvm-core/src/main/java/stryker4jvm/core/model/AST.java
@@ -1,4 +1,4 @@
-package core.model;
+package stryker4jvm.core.model;
 
 public abstract class AST {
     public abstract String syntax();
