+++ stryker4jvm-core/src/main/java/stryker4jvm/core/model/CollectedMutants.java
@@ -1,6 +1,7 @@
-package core.model;
+package stryker4jvm.core.model;
 
 import stryker4jvm.model.IgnoredMutationReason;
+import stryker4jvm.model.MutatedCode;
 
 import java.util.HashMap;
 import java.util.LinkedList;
@@ -9,9 +10,9 @@ import java.util.Map;
 
 public class CollectedMutants<T extends AST> {
     public final List<IgnoredMutation<T>> ignoredMutations;
-    public final Map<T, List<MutatedCode<T>>> mutations;
+    public final Map<T, List<stryker4jvm.model.MutatedCode<T>>> mutations;
 
-    public CollectedMutants(List<IgnoredMutation<T>> ignoredMutations, Map<T, List<MutatedCode<T>>> mutations) {
+    public CollectedMutants(List<IgnoredMutation<T>> ignoredMutations, Map<T, List<stryker4jvm.model.MutatedCode<T>>> mutations) {
         this.ignoredMutations = ignoredMutations;
         this.mutations = mutations;
     }
