+++ stryker4jvm-core/src/main/java/stryker4jvm/core/model/Collector.java
@@ -1,4 +1,4 @@
-package core.model;
+package stryker4jvm.core.model;
 
 public interface Collector<T extends AST> {
     CollectedMutants<T> collect(T tree);
