+++ stryker4jvm-core/src/main/java/stryker4jvm/core/model/MutantMetaData.java
@@ -1,4 +1,4 @@
-package core.model;
+package stryker4jvm.core.model;
 
 import mutationtesting.Location;
