+++ stryker4jvm-core/src/main/java/stryker4jvm/core/model/MutatedCode.java
@@ -1,4 +1,5 @@
-package core.model;
+package stryker4jvm.core.model;
 
 public record MutatedCode<T>(T mutatedStatement, MutantMetaData metaData) {
+
 }
