+++ stryker4jvm-core/src/main/java/stryker4jvm/core/model/Parser.java
@@ -1,4 +1,4 @@
-package core.model;
+package stryker4jvm.core.model;
 
 import java.nio.file.Path;
