+++ stryker4jvm-core/src/main/scala/stryker4jvm/model/Mutant.scala
@@ -1,6 +1,7 @@
 package stryker4jvm.model
 
 import mutationtesting.{Location, MutantResult, MutantStatus}
+import stryker4jvm.core.model.MutatedCode
 
 final case class MutantId(value: Int) extends AnyVal {
   override def toString(): String = value.toString
