+++ stryker4jvm-mutator-kotlin/src/main/kotlin/stryker4jvm/mutator/kotlin/KotlinAST.kt
@@ -1,7 +1,6 @@
 package stryker4jvm.mutator.kotlin
 
 import org.jetbrains.kotlin.psi.KtElement
-import stryker4jvm.mutants.language.AST
 
 class KotlinAST(val tree: KtElement) : AST() {
