+++ stryker4jvm-mutator-kotlin/src/main/kotlin/stryker4jvm/mutator/kotlin/KotlinInstrumenter.kt
@@ -1,8 +1,7 @@
 package stryker4jvm.mutator.kotlin
 
-import org.jetbrains.kotlin.psi.KtElement
 import scala.collection.Seq
-import stryker4jvm.model.MutantWithId
+import stryker4jvm.core.model.MutantWithId
 import stryker4jvm.mutants.language.Instrumenter
 
 class KotlinInstrumenter : Instrumenter<KotlinAST> {
