+++ stryker4jvm-mutator-scala/src/main/scala/stryker4jvm/mutants/Mutator.scala
@@ -9,6 +9,7 @@ import fs2.io.file.Path
 import fs2.{Chunk, Pipe, Stream}
 import mutationtesting.{MutantResult, MutantStatus}
 import stryker4jvm.config.Config
+import stryker4jvm.core.model
 import stryker4jvm.logging.Logger
 import stryker4jvm.model.*
 import stryker4jvm.mutants.findmutants.MutantFinder
@@ -26,7 +27,7 @@ class Mutator(
 ) {
 
   type Found[A, B] = (SourceContext, (Vector[A], Map[PlaceableTree, B]))
-  type FoundMutations = Found[(MutatedCode, IgnoredMutationReason), Mutations]
+  type FoundMutations = Found[(model.MutatedCode, IgnoredMutationReason), Mutations]
   type FoundMutationsWithId = Found[MutantResult, MutantsWithId]
 
   def go(files: Stream[IO, Path]): IO[(MutantResultsPerFile, Seq[MutatedFile])] = {
@@ -57,7 +58,7 @@ class Mutator(
 
   private def updateWithId: Pipe[IO, FoundMutations, FoundMutationsWithId] = {
 
-    def mapLeft(lefts: Vector[(MutatedCode, IgnoredMutationReason)], i: AtomicInteger) =
+    def mapLeft(lefts: Vector[(model.MutatedCode, IgnoredMutationReason)], i: AtomicInteger) =
       lefts.map { case (mutated, reason) =>
         MutantResult(
           i.getAndIncrement().toString(),
