+++ stryker4jvm-mutator-scala/src/main/scala/stryker4jvm/mutants/tree/MutantCollector.scala
@@ -1,8 +1,9 @@
 package stryker4jvm.mutants.tree
 
 import cats.syntax.align.*
+import stryker4jvm.core.model.MutatedCode
 import stryker4jvm.extensions.TreeExtensions.*
-import stryker4jvm.model.{IgnoredMutationReason, MutatedCode, PlaceableTree}
+import stryker4jvm.model.{IgnoredMutationReason, PlaceableTree}
 import stryker4jvm.mutants.Traverser
 import stryker4jvm.mutants.findmutants.MutantMatcher
