+++ stryker4jvm-mutator-scala/src/main/scala/stryker4jvm/mutants/tree/MutantInstrumenter.scala
@@ -2,10 +2,11 @@ package stryker4jvm.mutants.tree
 
 import cats.data.{NonEmptyList, NonEmptyVector}
 import cats.syntax.all.*
+import stryker4jvm.core.model.MutantWithId
 import stryker4jvm.extensions.TreeExtensions.TransformOnceExtension
 import stryker4jvm.exception.{Stryker4sException, UnableToBuildPatternMatchException}
 import stryker4jvm.logging.Logger
-import stryker4jvm.model.{MutantId, MutantWithId, MutatedFile, PlaceableTree, SourceContext}
+import stryker4jvm.model.{MutantId, MutatedFile, PlaceableTree, SourceContext}
 
 import scala.meta.*
 import scala.util.control.NonFatal
