+++ stryker4jvm-mutator-scala/src/test/scala/stryker4jvm/mutants/tree/MutantCollectorTest.scala
@@ -5,7 +5,7 @@ import cats.syntax.either.*
 import cats.syntax.option.*
 import mutationtesting.{Location, Position}
 import stryker4jvm.config.Config
-import stryker4jvm.model.{MutantMetadata, MutatedCode, PlaceableTree}
+import stryker4jvm.model.{MutantMetadata, PlaceableTree}
 import stryker4jvm.mutants.findmutants.{MutantMatcher, MutantMatcherImpl}
 import stryker4jvm.mutants.{Traverser, TraverserImpl}
 import stryker4jvm.scalatest.LogMatchers
