+++ stryker4jvm/src/main/scala/stryker4jvm/run/MutantRunner.scala
@@ -4,22 +4,15 @@ import cats.data.{EitherT, NonEmptyList}
 import cats.effect.{IO, Resource}
 import cats.syntax.all.*
 import fs2.io.file.{Files, Path}
-import fs2.{text, Pipe, Stream}
+import fs2.{Pipe, Stream, text}
 import mutationtesting.{MutantResult, MutantStatus}
 import stryker4jvm.config.Config
+import stryker4jvm.core.model.MutantWithId
 import stryker4jvm.exception.{InitialTestRunFailedException, UnableToFixCompilerErrorsException}
 import stryker4jvm.extensions.FileExtensions.PathExtensions
 import stryker4jvm.files.FilesFileResolver
 import stryker4jvm.logging.Logger
-import stryker4jvm.model.{
-  CompilerErrMsg,
-  InitialTestRunCoverageReport,
-  MutantResultsPerFile,
-  MutantWithId,
-  MutatedFile,
-  NoCoverageInitialTestRun,
-  RunResult
-}
+import stryker4jvm.model.{CompilerErrMsg, InitialTestRunCoverageReport, MutantResultsPerFile, MutatedFile, NoCoverageInitialTestRun, RunResult}
 import stryker4jvm.reporting.{MutantTestedEvent, Reporter}
 
 import java.nio
